Refactor - move UserStore out to its own class
