+++ src/app/users.spec.ts
@@ -1,0 +1,7 @@
+import { Users } from './users';
+
+describe('Users', () => {
+  it('should create an instance', () => {
+    expect(new Users()).toBeTruthy();
+  });
+});
